% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{capitalize_prefix}
\alias{capitalize_prefix}
\title{Capitalizes all letters in a character string until a separator character}
\usage{
capitalize_prefix(char, sep = "_")
}
\arguments{
\item{char}{A character string.}

\item{sep}{The separator character.}
}
\value{
The modified character string.
}
\description{
Returns original character string if it does not contain the separator
character.
}
\examples{
char <- "adsl_depression"
capitalize_prefix(char)
}
