% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mode}
\alias{mode}
\title{Get the mode of a vector.}
\usage{
mode(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The mode of \code{x}.
}
\description{
The mode of a vector is the value that appears most often.
}
\examples{
mode(c("D", "A", "B", "B", "C"))


# In case of 2 or more modi, the function returns the modus that appears
# first:
mode(c(4, 9, 1, 2, 9, 1))
}
